feat: add different modes for horizontal scaling
